Refactor search-adapter.ts code and split it into multiple files
