+++ packages/framework-provider-aws/src/library/read-models-adapter.ts
